+++ server/src/test/java/org/elasticsearch/discovery/zen/MembershipActionTests.java
@@ -19,6 +19,7 @@
 package org.elasticsearch.discovery.zen;
 
 import org.elasticsearch.Version;
+import org.elasticsearch.cluster.coordination.JoinTaskExecutor;
 import org.elasticsearch.cluster.metadata.IndexMetaData;
 import org.elasticsearch.cluster.metadata.MetaData;
 import org.elasticsearch.cluster.node.DiscoveryNode;
@@ -49,10 +50,10 @@ public class MembershipActionTests extends ESTestCase {
             .numberOfReplicas(1).build();
         metaBuilder.put(indexMetaData, false);
         MetaData metaData = metaBuilder.build();
-        MembershipAction.ensureIndexCompatibility(Version.CURRENT, metaData);
+        JoinTaskExecutor.ensureIndexCompatibility(Version.CURRENT, metaData);
 
         expectThrows(IllegalStateException.class, () ->
-        MembershipAction.ensureIndexCompatibility(VersionUtils.getPreviousVersion(Version.CURRENT),
+        JoinTaskExecutor.ensureIndexCompatibility(VersionUtils.getPreviousVersion(Version.CURRENT),
             metaData));
     }
 
@@ -67,7 +68,7 @@ public class MembershipActionTests extends ESTestCase {
         metaBuilder.put(indexMetaData, false);
         MetaData metaData = metaBuilder.build();
         expectThrows(IllegalStateException.class, () ->
-            MembershipAction.ensureIndexCompatibility(Version.CURRENT,
+            JoinTaskExecutor.ensureIndexCompatibility(Version.CURRENT,
                 metaData));
     }
 
@@ -84,9 +85,9 @@ public class MembershipActionTests extends ESTestCase {
             final Version tooLow = getPreviousVersion(maxNodeVersion.minimumCompatibilityVersion());
             expectThrows(IllegalStateException.class, () -> {
                 if (randomBoolean()) {
-                    MembershipAction.ensureNodesCompatibility(tooLow, nodes);
+                    JoinTaskExecutor.ensureNodesCompatibility(tooLow, nodes);
                 } else {
-                    MembershipAction.ensureNodesCompatibility(tooLow, minNodeVersion, maxNodeVersion);
+                    JoinTaskExecutor.ensureNodesCompatibility(tooLow, minNodeVersion, maxNodeVersion);
                 }
             });
         }
@@ -95,16 +96,16 @@ public class MembershipActionTests extends ESTestCase {
             Version tooHigh = incompatibleFutureVersion(minNodeVersion);
             expectThrows(IllegalStateException.class, () -> {
                 if (randomBoolean()) {
-                    MembershipAction.ensureNodesCompatibility(tooHigh, nodes);
+                    JoinTaskExecutor.ensureNodesCompatibility(tooHigh, nodes);
                 } else {
-                    MembershipAction.ensureNodesCompatibility(tooHigh, minNodeVersion, maxNodeVersion);
+                    JoinTaskExecutor.ensureNodesCompatibility(tooHigh, minNodeVersion, maxNodeVersion);
                 }
             });
         }
 
         if (minNodeVersion.onOrAfter(Version.V_6_0_0_alpha1)) {
             Version oldMajor = randomFrom(allVersions().stream().filter(v -> v.major < 6).collect(Collectors.toList()));
-            expectThrows(IllegalStateException.class, () -> MembershipAction.ensureMajorVersionBarrier(oldMajor, minNodeVersion));
+            expectThrows(IllegalStateException.class, () -> JoinTaskExecutor.ensureMajorVersionBarrier(oldMajor, minNodeVersion));
         }
 
         final Version minGoodVersion = maxNodeVersion.major == minNodeVersion.major ?
@@ -114,9 +115,9 @@ public class MembershipActionTests extends ESTestCase {
         final Version justGood = randomVersionBetween(random(), minGoodVersion, maxCompatibleVersion(minNodeVersion));
 
         if (randomBoolean()) {
-            MembershipAction.ensureNodesCompatibility(justGood, nodes);
+            JoinTaskExecutor.ensureNodesCompatibility(justGood, nodes);
         } else {
-            MembershipAction.ensureNodesCompatibility(justGood, minNodeVersion, maxNodeVersion);
+            JoinTaskExecutor.ensureNodesCompatibility(justGood, minNodeVersion, maxNodeVersion);
         }
     }
 
@@ -136,7 +137,7 @@ public class MembershipActionTests extends ESTestCase {
             .numberOfReplicas(1).build();
         metaBuilder.put(indexMetaData, false);
         MetaData metaData = metaBuilder.build();
-            MembershipAction.ensureIndexCompatibility(Version.CURRENT,
+            JoinTaskExecutor.ensureIndexCompatibility(Version.CURRENT,
                 metaData);
     }
 }
